refactor: modify logo image & update imports
